ToBeContinued

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: mbuchet <mbuchet@student.s19.be>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/01 16:10:40 by mbuchet           #+#    #+#             */
-/*   Updated: 2022/02/02 21:28:07 by mbuchet          ###   ########.fr       */
+/*   Updated: 2022/02/03 16:13:02 by mbuchet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_printf.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != 0)
